Casted Math.ceil to int and gave Horse saddle

+++ src/net/komputerking/java/tendotjava/Main.java
@@ -35,6 +35,9 @@ public class Main extends JavaPlugin{
     public void onDisable(){
     }
     
+    /**
+     * Drops a crate at a random location in a loaded chunk.
+     */
     public void dropCrate(){
         if (found){
         Random rand = new Random();
@@ -43,7 +46,7 @@ public class Main extends JavaPlugin{
         int chunkSeed = rand.nextInt(w.getLoadedChunks().length);
         Chunk c = w.getLoadedChunks()[chunkSeed];
         final Location loc = new Location(w, c.getX(), 255, c.getZ());
-        Bukkit.broadcastMessage(ChatColor.GRAY + "[" + ChatColor.GOLD + "EntityCrates" + ChatColor.GRAY + "]" + ChatColor.WHITE + " A crate is dropping at the highest point at " + Math.ceil(loc.getX()) + ", " + Math.ceil(loc.getZ()));
+        Bukkit.broadcastMessage(ChatColor.GRAY + "[" + ChatColor.GOLD + "EntityCrates" + ChatColor.GRAY + "]" + ChatColor.WHITE + " A crate is dropping at the highest point at " + (int) Math.ceil(loc.getX()) + ", " + (int) Math.ceil(loc.getZ()));
         Bukkit.broadcastMessage(ChatColor.GRAY + "[" + ChatColor.GOLD + "EntityCrates" + ChatColor.GRAY + "]" + ChatColor.WHITE + " Get there to pick it up.");
         Bukkit.getScheduler().scheduleSyncDelayedTask(this, new Runnable(){
             public void run(){
